relcoate the filter to be loaded in rest as well
